Moved SearchInput component in main component directory
