+++ src/screens/SearchScreen.js
@@ -5,8 +5,8 @@ import { Icon } from "react-native-elements";
 import { ScrollView } from "react-native-gesture-handler";
 import { useDispatch, useSelector } from "react-redux";
 import tw from "tailwind-react-native-classnames";
+import SearchInput from "../components/SearchInput";
 import StackBottomSheet from "../components/StackBottomSheet";
-import SearchInput from "../components/subcomponents/SearchInput";
 import { selectNearbyPlaces, selectUserLocation } from "../features/locationSlice";
 import { setOrigin, selectOrigin } from "../features/navSlice";
 import { textEllipsis, randomizeArray } from "../Helper";
